refactor: monorepo setup
