+++ lerna.json
@@ -1,0 +1,21 @@
+{
+	"packages": [
+		"storages/*",
+		"ts-cache"
+	],
+	"version": "independent",
+	"command": {
+		"publish": {
+			"message": "chore(release): publish"
+		}
+	},
+	"hoist": false,
+	"forceLocal": true,
+	"conventionalCommits": true,
+	"reject-cycles": true,
+	"ignoreChanges": [
+		"**/__tests__/**",
+		"**/*.md",
+		"**/*.tmp"
+	]
+}
