+++ storages/fs/src/index.ts
@@ -1,0 +1,1 @@
+export { FsJsonStorage } from './fs.json.storage';
