+++ storages/node-cache/src/index.ts
@@ -1,0 +1,1 @@
+export { NodeCacheStorage } from './node-cache.storage';
