+++ storages/node-cache/src/node-cache.storage.ts
@@ -1,4 +1,4 @@
-import { StorageTypes } from './storage.types'
+import { StorageTypes } from 'node-ts-cache'
 
 import * as NodeCache from  "node-cache" ;
