+++ storages/redis/src/index.ts
@@ -1,0 +1,1 @@
+export { RedisStorage } from './redis.storage';
