+++ test/memory.storage.test.ts
@@ -1,5 +1,5 @@
 import * as Assert from 'assert'
-import { MemoryStorage } from '../src'
+import { MemoryStorage } from '../ts-cache/src'
 
 describe('MemoryStorage', () => {
     it('Should add cache item correctly', async () => {
