+++ test/redis.storage.test.ts
@@ -1,6 +1,6 @@
 import * as Proxyquire from 'proxyquire'
 import * as Assert from 'assert'
-import { RedisStorage } from '../src'
+import { RedisStorage } from '../ts-cache/src'
 
 Proxyquire.noCallThru()
