+++ ts-cache/src/index.ts
@@ -1,0 +1,5 @@
+export { StorageTypes } from './storage/storage.types'
+export { ExpirationStrategy } from './strategy/caching/expiration.strategy'
+export { IKeyStrategy } from './strategy/key/key.strategy.types'
+export { ICacheStrategy } from './strategy/caching/cache.strategy.types'
+export { Cache } from './decorator/cache.decorator'
