+++ ts-cache/src/storage/storage.types.ts
